moved open_url to a utilities folder so it could be used across plugin; moved auto_expand functions to goodreads parser so it can be used during xray creation as well as book specific preferences
