+++ lib/exceptions.py
@@ -1,0 +1,6 @@
+# exceptions.py
+'''Class of custom exceptions'''
+
+class PageDoesNotExist(Exception):
+    '''Exception for when page does not exist'''
+    pass
